pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mcserver-status-check"
version = "0.1.0"
description = "一款原版风格、美观高效的 Minecraft 服务器状态查询插件"
authors = ["leiuary <leiuary@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/leiuary/nonebot-plugin-mcserver-status-check"
repository = "https://github.com/leiuary/nonebot-plugin-mcserver-status-check"
keywords = ["nonebot", "nonebot2", "minecraft", "mcstatus"]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = "^2.0.0"
Pillow = "^10.0.0"
mcstatus = "^11.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
